# api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateFacultyDTO:
    properties:
      is_active:
        type: boolean
      name_en:
        type: string
      name_fa:
        type: string
      short_code:
        maxLength: 10
        type: string
      university_id:
        type: string
    required:
    - name_en
    - name_fa
    - short_code
    - university_id
    type: object
  dto.CreateProfessorRequest:
    properties:
      name:
        type: string
      university_id:
        type: string
    required:
    - name
    - university_id
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        description: |-
          error message
          example: invalid request format
        type: string
    type: object
  dto.FacultyResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name_en:
        type: string
      name_fa:
        type: string
      short_code:
        type: string
      university_id:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.ProfessorResponse:
    properties:
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      normalizedName:
        type: string
      universityID:
        type: string
      updatedAt:
        type: string
    type: object
...
  dto.UpdateFacultyDTO:
    properties:
      is_active:
        type: boolean
      name_en:
        type: string
      name_fa:
        type: string
      short_code:
        maxLength: 10
        type: string
      university_id:
        type: string
    required:
    - name_en
    - name_fa
    - short_code
    - university_id
    type: object
...
host: localhost:8080
info:
  contact: {}
  title: Termustat API
  version: "1.0"
paths:
...
  /v1/admin/faculties:
    post:
      consumes:
      - application/json
      description: Creates a new faculty under a university
      parameters:
      - description: Create faculty payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateFacultyDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.FacultyResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: University not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict (e.g. short code exists)
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Failed to create faculty
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Faculty
      tags:
      - faculties
  /v1/admin/faculties/{id}:
    delete:
      description: Deletes a faculty from the system
      parameters:
      - description: Faculty ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Faculty deleted successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid faculty ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Faculty not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Failed to delete faculty
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Faculty
      tags:
      - faculties
    get:
      description: Retrieves a faculty by its unique ID
      parameters:
      - description: Faculty ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FacultyResponse'
        "400":
          description: Invalid faculty ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Faculty not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Failed to get faculty
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Faculty
      tags:
      - faculties
    put:
      consumes:
      - application/json
      description: Updates the details of an existing faculty
      parameters:
      - description: Faculty ID
        in: path
        name: id
        required: true
        type: string
      - description: Update faculty payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateFacultyDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FacultyResponse'
        "400":
          description: Invalid request body or ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Faculty not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict updating faculty
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Failed to update faculty
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Faculty
      tags:
      - faculties
  /v1/admin/professors:
    post:
      consumes:
      - application/json
      description: Creates a new professor under a university, or returns existing
        one
      parameters:
      - description: Create professor payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProfessorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ProfessorResponse'
        "400":
          description: Invalid payload or university not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Professor
      tags:
      - professors
  /v1/admin/professors/{id}:
    get:
      description: Retrieves a professor by their unique ID
      parameters:
      - description: Professor ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProfessorResponse'
        "400":
          description: Invalid professor ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Professor not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Professor
      tags:
      - professors
...
  /v1/admin/universities/{id}/faculties:
    get:
      description: Retrieves all faculties under the specified university
      parameters:
      - description: University ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.FacultyResponse'
            type: array
        "400":
          description: Invalid university ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: University not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Failed to get faculties
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List Faculties
      tags:
      - faculties
  /v1/admin/universities/{id}/faculty/{short_code}:
    get:
      description: Retrieves a faculty within a university by its short code
      parameters:
      - description: University ID
        in: path
        name: id
        required: true
        type: string
      - description: Faculty short code
        in: path
        name: short_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FacultyResponse'
        "400":
          description: Invalid university ID or short code missing
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Faculty not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Failed to get faculty
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Faculty by Short Code
      tags:
      - faculties
  /v1/admin/universities/{id}/professors:
    get:
      description: Retrieves all professors associated with a given university
      parameters:
      - description: University ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ProfessorResponse'
            type: array
        "400":
          description: Invalid university ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List Professors
      tags:
      - professors
...
swagger: "2.0"